server/requirements.txt
Flask==3.1.3
flask-cors==5.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
